requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
PyMySQL==1.0.2
Jinja2==2.11.3
Werkzeug==1.0.1
pdf2image==1.14.0
wtforms==2.3.3
pykakasi==2.0.6
numpy==1.16.4
pillow==8.1.0
opencv-python==4.4.0.46
matplotlib==3.3.3
pandas==1.1.5
h5py==2.10.0
SQLAlchemy==1.3.23
mojimoji==0.0.11